pyproject.toml
[project]
name = "aws-native-rag-service"
version = "0.1.0"
description = "Production-grade AWS-native Retrieval Augmented Generation service"
requires-python = ">=3.11"
dependencies = [
    "aws-cdk-lib>=2.170.0",
    "constructs>=10.0.0",
    "boto3>=1.35.0",
    "opensearch-py>=2.4.0",
    "redis>=5.0.0",
    "aws-xray-sdk>=2.12.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "hypothesis>=6.100.0",
    "moto[all]>=5.0.0",
    "fakeredis>=2.21.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "ragas>=0.1.0",
    "boto3-stubs[essential]>=1.35.0",
    "requests>=2.31.0",
]

[tool.black]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
markers = [
    "integration: marks tests as integration tests (require deployed environment)",
]

[tool.mypy]
strict = true
